[ADD] most the application finished

+++ js/controllers/myFavoritesController.js
@@ -2,6 +2,8 @@ talenting.controller('myFavoritesController', ['$scope', '$http', '$location','u
 
     $scope.userSession;
     $scope.myFavoriteOnes = [];
+    $scope.vacancies = {};
+    $scope.retrievedBenefits = [];
 
     let session = $cookies.get('user');
 
@@ -117,15 +119,36 @@ talenting.controller('myFavoritesController', ['$scope', '$http', '$location','u
     // to disapple to a vacancy
     $scope.disapply = (vacancyId, index) => {
 
-        $http({
-            method: "DELETE",
-            url: 'http://localhost:8080/talenting/appliersInVacancies',
-            params: {vacancyId: vacancyId}
-        }).then( () => {
+        Swal.fire({
+            title: '¿Estas seguro?',
+            text: "Si retiras tu postulación todo tu proceso de contratación será cancelado",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#198754',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $http({
+                    method: "DELETE",
+                    url: 'http://localhost:8080/talenting/appliersInVacancies',
+                    params: {vacancyId: vacancyId}
+                }).then( () => {
+                    $scope.myFavoriteOnes[index].applied = 0;
+                    alertService.showAlert.info('¡Se ha retirado tu postulación!');
+                });
+            }
+        })
+    };
 
-            $scope.myFavoriteOnes[index].applied = 0;
-            
-        });
+
+    $scope.openModalInformation = (fullVacancy) => {
+
+        console.log(fullVacancy);
+        $scope.vacancies = angular.copy(fullVacancy);
+        $("#infoPostulation").modal("show");
+
     };
 
 }]);
